requirements.txt
pandas<1.3.4
Pillow==9.0.0
plotly==5.5.0
streamlit==1.5.0
termcolor==1.1.0
openpyxl<3.0.9



